requirements.txt
torch>=2.0.0
numpy
pyyaml
adam-atan2
einops
flash-attn
